+++ .gitignore
@@ -19,6 +19,8 @@
 *.tar.gz
 *.rar
 
+src/main/java/Api/Credentials.java
+
 src/main/resources/*.json
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
