+++ src/main/java/UseCase/AvailableContest/AvailableCOntestOutputData.java
@@ -1,0 +1,4 @@
+package UseCase.AvailableContest;
+
+public class AvailableCOntestOutputData {
+}
