+++ src/main/java/UseCase/AvailableContest/AvailableContestInputData.java
@@ -1,0 +1,4 @@
+package UseCase.AvailableContest;
+
+public class AvailableContestInputData {
+}
