+++ src/tests/java/AvailableContestsTests/CodeBehindTests/AvailableContestInteractorTest.java
@@ -1,0 +1,20 @@
+package java.AvailableContestsTests.CodeBehindTests;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class AvailableContestInteractorTest {
+
+    @Test
+    void execute() {
+    }
+
+    @Test
+    void enrollUserInContest() {
+    }
+
+    @Test
+    void getUpdatedStockPrice() {
+    }
+}
