+++ src/tests/java/EnrolledContestTests/EnrolledContestViewMain.java
@@ -1,12 +1,9 @@
-package main.tests.EnrolledContestTests;
+package java.EnrolledContestTests;
 
 import FirebaseDataAccess.FirebaseDataAccess;
 import app.ContestUseCaseFactory;
-import com.google.auth.oauth2.GoogleCredentials;
+import app.Main;
 import com.google.cloud.firestore.Firestore;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseOptions;
-import com.google.firebase.cloud.FirestoreClient;
 import entity.Contest;
 import interface_adapters.Enrolled.EnrolledViewModel;
 import interface_adapters.ViewModelManager;
@@ -14,38 +11,21 @@ import view.EnrolledContest.EnrolledView;
 
 import javax.swing.*;
 import java.awt.*;
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.net.URL;
 
-public class EnrolledUseCaseTest {
+public class EnrolledContestViewMain {
+
+
     public static void main(String[] args) {
         // initialization for firebase.
         Firestore db;
 
         try{
-
-            URL url =  main.tests.EnrolledContestTests.EnrolledContestViewMain.class.getResource("/wallstreetwarriors-firebase-adminsdk-8g503-275acc4c97.json");
-            File file = new File(url.getPath());
-
-            FileInputStream serviceAccount =
-                    new FileInputStream(file);
-
-            FirebaseOptions options = new FirebaseOptions.Builder()
-                    .setCredentials(GoogleCredentials.fromStream(serviceAccount))
-                    .build();
-
-            FirebaseApp.initializeApp(options);
-
-            db = FirestoreClient.getFirestore();
-
-            //"Initialize" singleton entity level data access factory
+            Main.FirebaseInit();
             var firebaseDataAccess = FirebaseDataAccess.getInstance();
-            firebaseDataAccess.setFirestore(db);
             // catching exceptions for GoogleCredentials.fromStream and FileInputStream
-            String contestId = "1";
+            String contestId = "1000";
             firebaseDataAccess.getEntity(Contest.class, "Contests", contestId);
 
             // creating JFrame for app
